refactored user services

+++ src/modules/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { UpdateUserDto } from './dto/user.dto';
 
@@ -11,14 +11,30 @@ export class UsersService {
     }
 
     async findOne(id: number) {
-        return await this.prisma.user.findUnique({ where: { id } });
+        const existingUser = await this.prisma.user.findUnique({ where: { id } });
+
+        if (!existingUser) {
+            throw new NotFoundException('User not found');
+        }
+        return existingUser;
     }
 
     async update(id: number, updateUserData: UpdateUserDto) {
+        const existingUser = await this.prisma.user.findUnique({ where: { id } });
+
+        if (!existingUser) {
+            throw new NotFoundException('User not found');
+        }
+
         return await this.prisma.user.update({ where: { id }, data: updateUserData });
     }
 
     async remove(id: number) {
+        const existingUser = await this.prisma.user.findUnique({ where: { id } });
+
+        if (!existingUser) {
+            throw new NotFoundException('User not found');
+        }
         return await this.prisma.user.delete({ where: { id } });
     }
 }
